POC of pydantic-to-typescript for generating frontend types.
