+++ requirements.txt
@@ -83,3 +83,4 @@ urllib3==2.2.2
 wrapt==1.16.0
 zipp==3.19.2
 EDMesg @ git+https://github.com/RatherRude/EDMesg.git@e4a387e
+pydantic-to-typescript==2.0.0
