+++ ui/package.json
@@ -4,10 +4,11 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "dev": "tauri dev",
-    "build": "ng build",
-    "watch": "ng build --watch --configuration development",
-    "tauri": "tauri"
+    "generate-types": "pydantic2ts --module ../src/lib/shared/models.py --output src/app/types/models.d.ts",
+    "dev": "npm run generate-types && tauri dev",
+    "build": "npm run generate-types && ng build",
+    "watch": "npm run build --watch --configuration development",
+    "tauri": "npm run generate-types && tauri"
   },
   "private": true,
   "dependencies": {
@@ -35,6 +36,7 @@
     "@tauri-apps/cli": "^2",
     "@types/jasmine": "~5.1.0",
     "jasmine-core": "~5.1.0",
+    "json-schema-to-typescript": "^15.0.4",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
