+++ ui/src/app/types/models.d.ts
@@ -1,0 +1,13 @@
+/* tslint:disable */
+/* eslint-disable */
+/**
+/* This file was automatically generated from pydantic models by running pydantic2ts.
+/* Do not modify it by hand - just update the pydantic models and then re-run the script
+*/
+
+export interface SystemInfo {
+  os: string;
+  input_device_names: string[];
+  output_device_names: string[];
+  edcopilot_installed: boolean;
+}
